requirements.txt
tkinter
matplotlib
pandas
sklearn
MySQL.connector